Fix wrong child representation returned by toTree
After checking if there are multiple children (array of children should stay flat)
check if the child is a wrapper for Text Node.
If it is parse the child text node and wrap it in an array (but only if a valid child is returned).

Includes fix of test for master: due to merge